+++ src/components/SponsorsPageInside/Left/InsideLP.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styles from './InsideL.module.css'
-import SmallBox from './Polaroid.jsx' 
+import SmallBox from '../Polaroid/Polaroid.jsx' 
 
 const FilesLeftTitle = () => {
   return (
@@ -11,11 +11,13 @@ const FilesLeftTitle = () => {
 
                 <div className={styles.filespageL2}>
                 <h1 classname={styles.invi}>&nbsp;</h1>
-                <h1 align="center">TITLE</h1>
+                <h1 align="center">POWER</h1>
                 <img className={styles.SpyGunLogo}  src="public\Sponsors\ClosedFile\007-Logo.png"  alt="007Logo"/>
                 <img className={styles.target}      src="public\Sponsors\ClosedFile\Crosshair.png"  alt="Crosshair"/>
+                <div className={styles.center}><SmallBox/><SmallBox/></div>
                 <div className={styles.center}><SmallBox/></div>
-
+                <div className={styles.center}><SmallBox/><SmallBox/></div>
+                
                 </div>
 
             </div>
