+++ src/components/SponsorsPageInside/PolaroidsSettings/Title.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import styles from './InsideR.module.css'
-import SmallBox from './Polaroid.jsx' 
+import styles from '../Right/InsideR.module.css'
+import SmallBox from '../Polaroid/Polaroid.jsx' 
 
 const FilesPower = () => {
   return (
