+++ src/components/SponsorsPageInside/Right/InsideRA.jsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import styles from './InsideR.module.css'
-import Title from './Title.jsx'
+import Associate from '../PolaroidsSettings/Associate.jsx'
 
-const FileTitle = () => {
+const FileAssociate = () => {
   return (
     <>
         <div className={styles.filespagebgR}>
             <div className={styles.filespageR1}>
                 <div className={styles.filespageR2}>
                     <img className={styles.handprint}   src="public\Sponsors\ClosedFile\handprint.png"  alt="HandPrint"/>
-                    <Title />
+                    <Associate />
                     <img className={styles.handcuffs}   src="public\Sponsors\ClosedFile\Hatkadi.png"    alt="Handcuffs"/>
                 </div>
             </div>
@@ -17,5 +17,5 @@ const FileTitle = () => {
     </>
 )
 }
-
-export default FileTitle
+    
+export default FileAssociate
