+++ src/components/SponsorsPageInside/Right/InsideRP.jsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import styles from './InsideR.module.css'
-import Associate from './Associate.jsx'
+import Power from '../PolaroidsSettings/Power.jsx'
 
-const FileAssociate = () => {
+const FileTitle = () => {
   return (
     <>
         <div className={styles.filespagebgR}>
             <div className={styles.filespageR1}>
                 <div className={styles.filespageR2}>
                     <img className={styles.handprint}   src="public\Sponsors\ClosedFile\handprint.png"  alt="HandPrint"/>
-                    <Associate />
+                    <Power />
                     <img className={styles.handcuffs}   src="public\Sponsors\ClosedFile\Hatkadi.png"    alt="Handcuffs"/>
                 </div>
             </div>
@@ -17,5 +17,5 @@ const FileAssociate = () => {
     </>
 )
 }
-    
-export default FileAssociate
+
+export default FileTitle
